update: split deploy
